requirements.txt
pyfiglet
python-Wappalyzer